Refactor code architecture with service layer and contracts
- Extract tool management into AgentToolService class
- Add ToolServiceContract interface for better abstraction
- Move file utilities to separate utils.py module
- Remove tools.py and restructure tool handling
- Update Agent class to use dependency injection pattern
- Improve separation of concerns and code organization
