+++ projects/1-simple-file-agent/run.py
@@ -1,8 +1,8 @@
 from agent import Agent
-from tools import tools_definition
+from agent_tool_service import AgentToolService
 
 task: str = input('Hi there! How can I help you? \nYou>> ')
 # message: str = "What's inside my current folder?"
 
-agent: Agent = Agent(tools_definition)
+agent: Agent = Agent(AgentToolService())
 agent.run(task)
