+++ projects/1-simple-file-agent/utils.py
@@ -1,0 +1,3 @@
+def read_file(path: str) -> str:
+    with open(path, 'r') as f:
+        return f.read()
